pyproject.toml
[project]
name = "substack-mcp"
version = "0.1.0"
description = "MCP server for Substack — create, update, schedule, publish posts with images and cover thumbnails"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "Taiyo Kimura" }]
dependencies = [
    "mcp>=1.2.0",
    "python-substack>=0.1.20",
    "pydantic>=2.0",
    "platformdirs>=4.0",
    "playwright>=1.40",
    "pycookiecheat>=0.7",
]

[project.scripts]
substack-mcp = "substack_mcp.server:main"
substack-mcp-setup = "substack_mcp.setup_cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/substack_mcp"]
